update exercise 7

+++ exercise7/internal/api/inputs/product_create_input.go
@@ -2,6 +2,6 @@ package inputs
 
 type CreateProductInput struct {
 	Name     string `json:"name" binding:"required"`
-	Price    int    `json:"price" binding:"required"`
+	Price    float64    `json:"price" binding:"required"`
 	Quantity int    `json:"quantity" binding:"required"`
 }
